Organize files better
First things first... organize!
