+++ plugins/UM3NetworkPrinting/__init__.py
@@ -1,11 +1,11 @@
 # Copyright (c) 2017 Ultimaker B.V.
 # Cura is released under the terms of the LGPLv3 or higher.
 
-from . import DiscoverUM3Action
+from .src import DiscoverUM3Action
 from UM.i18n import i18nCatalog
 catalog = i18nCatalog("cura")
 
-from . import UM3OutputDevicePlugin
+from .src import UM3OutputDevicePlugin
 
 def getMetaData():
     return {}
