+++ plugins/UM3NetworkPrinting/resources/qml/MonitorItem.qml
