+++ plugins/UM3NetworkPrinting/resources/qml/PrinterVideoStream.qml
